examples/genai_rag_solution/components/frontend/requirements.txt
streamlit==1.30.0
streamlit-chat==0.1.1
streamlit-javascript==0.1.5
streamlit-extras==0.3.6
streamlit-oauth==0.1.8
moment==0.12.1
python-dotenv==1.0.1
